runservice: add tasks groups
Task groups permit to group tasks together. Like tasks they can depend on each
other. A task group is finished when all the tasks inside a task group are
finished. This is required since we want to have reproducible workspaces.
A child task group will start when all the parent task groups are finished.

When a task in a child task group restores a workspace, it'll consider all the
tasks in parent task groups as a dependency of the parent task in this task
group.