*: replace etcd and datamanager with a sql db
Replace etcd and datamanager with a sql db (postgres or sqlite)
At the start of agola we used etcd and an object storage so we tried to
avoid the need to add an additional sql database layer by writing a
quite complex "datamanager" to store data and local sqlite databases
synced with it to query data. This was quite powerful but also very
complex to understand for new developers. So this patch removes it and
replace the datamanager and the use of etcd by using a sql database.
The objectstorage is used to store other data (like logs, archives
etc...) but not the runs data and the configstore data.

* Added db configuration entry for runservice, configstore and
  notification service (used only for locking).
* Added a generic sql helper package. The transaction isolation level
  required by the underlying db is always "serializable" to avoid any
  concurrency issue. Additionally we implement also optimistic locking
  to span multiple transactions using a revision column.
* Implement a generic db layer based on objects.
  * We don't rely on any specific database feature.
  * We have data tables that have only 3 columns: id, revision and data
    (containing the json marshalled object)
  * There can be multiple "query" tables that are kept synced (at the
    transaction level) with the data tables and are used for all the
    application logic.
  * The query tables can be dropped and rebuilt from the data tables
    when we need to change their schema or add more query tables.
  * We can have multiple query tables for the same data objects type to
    speed up things.
  * In future also synced materialized views could be implemented if
    there're performance reasons.
* Add a lock package to do optimistic locking using local locks (for
  sqlite) or postgres based locks.