gateway: use auth cookies instead of jwt tokens
The authentication won't return a jwt token but will use cookies.
The token auth will be used only for api calls with the admintoken or
user token.

Two cookies are returned, the primary auth cookie (httpOnly) and a
secondary cookie used for client side implementation of logout.

Also implement CSRF and use of secure cookies (with a config option to
disable them)