db: improve transactions handling
Since the transactions can be repeated (i.e. due to serializable errors
in postgres or locked errors in sqlite) we should ensure that the
objects are created inside the transaction or at the second execution
the transaction will fail since the object revision has been updated by
the previous one.

This patch adds a txID to every transaction that it's set on objects
fetched as an helper to ensure that the object is created/fetched inside
the same transaction.

To handle cases were the object is fetched/created outside the
transaction we MUST manually ensure that the object is still valid (the
tx logic must be self consistent) and only in such case restore the txID
to the current transactions and revision to the original value at the
start of the transaction.