tests: fix failing gitea pull request creation
Looks like there're some async handlings in gitea when pushing and then
instantly creating a pull request that could fail with a 404 Not found.

We tried with a wait, before pr creation, to check that the branch
exists but this always returns true also if the pr creation then fails.

So retry the creation for some time on 404.