gitsources: split oauth2Client from gitsource
Split an oauth2Client from the gitsource. While the gitsource creates a
client to the remote source an Oauth2Client doesn't requires a client to
the remote sources but only handles the oauth2 flow.

This also fixes an issue with gitea 1.14 (fixed in later gitea versions)
where the /version endpoint requires authentication but during the
oauth2 flow we don't have any auth data and so we cannot create the
gitea client because during gite client creation a call to the version
endpoint is done and returns an error.