configstore: add ability to org members to restart stop cancel a project run
This patch adds a project option to permit org members to execute run actions (restart/stop/cancel)
The `MembersCanPerformRunActions` option has been added to the project type.
The API will permit to set it to true on project create/update only for projects belonging to an organization and will return an error when on user's projects.