*: add detailed errors to api responses
In addition to the standard http error codes also report a detailed
error that contains an error code and an optional error detail. This is
required to better explain the reason of the reported error.

Since, by default if not handled in another way, an error returned by
an internal service is propagated by the gateway to the client, also
internal services returns detailed errors with exported error codes.

This patch adds an initial round of detailed errors. They are not stable
and will change in future.